refactor: clearness of folders
