+++ Assets/Dev/Arby/core_actions/Scripts/CoreActions/Input.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1cb39f8e9c4fe4946a0eb1a680adb815
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
