--- Assets/Dev/Arby/core_actions/Scripts/CoreActions/RobotActionInterface.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: a7f9adfd2929a48b6847ff65143f362d
